fix: fix org loading
