+++ src/pages/Background/modules/saveToGoDAM.js
@@ -1,144 +1,52 @@
-import signInGoDAM from "./signInGoDAM";
-
-// GoDAM OAuth configuration
-const clientId = process.env.GODAM_OAUTH_CLIENT_ID;
+import getGoDAMAuthToken from "./getGoDAMAuthToken";
 
 // Function to upload a video to GoDAM
 const saveToGoDAM = async (videoBlob, fileName, sendResponse) => {
-  // Function to get an access token from Chrome storage
-  async function getGoDAMAuthToken() {
-    return new Promise((resolve, reject) => {
-      chrome.storage.local.get(["godamToken", "godamRefreshToken", "godamTokenExpiration"], async (result) => {
-        if (chrome.runtime.lastError) {
-          reject(new Error(chrome.runtime.lastError));
-        } else {
-          const { godamToken, godamRefreshToken, godamTokenExpiration } = result;
-          
-          if (!godamToken) {
-            // Token is not set, trigger sign-in
-            const newToken = await signInGoDAM();
-            if (!newToken) {
-              // Sign-in failed, throw an error
-              reject(new Error("GoDAM sign-in failed"));
-            }
-            resolve(newToken);
-          } else {
-            // Token is set, check if it has expired
-            const currentTime = Date.now();
-            if (currentTime >= godamTokenExpiration) {
-              // Token has expired, refresh it
-              const baseUrl = process.env.GODAM_BASE_URL || 'https://app.godam.io';
-                try {
-                    const refreshResponse = await fetch(`${baseUrl}/api/method/frappe.integrations.oauth2.get_token`, {
-                        method: 'POST',
-                        headers: {
-                            'Content-Type': 'application/x-www-form-urlencoded',
-                            'Accept': 'application/json',
-                        },
-                        credentials: 'omit',
-                        body: new URLSearchParams({
-                            grant_type: 'authorization_code',
-                            refresh_token: godamRefreshToken,
-                            client_id: clientId,
-                        }),
-                    });
-
-                if (!refreshResponse.ok) {
-                  // Refresh token failed, try to sign in again
-                  const newToken = await signInGoDAM();
-                  if (!newToken) {
-                    reject(new Error("GoDAM token refresh and sign-in failed"));
-                  }
-                  resolve(newToken);
-                  return;
-                }
-
-                const tokenData = await refreshResponse.json();
-                const newToken = tokenData.access_token;
-                const expiresIn = tokenData.expires_in || 3600;
-                const expirationTime = Date.now() + expiresIn * 1000;
-
-                // Save new token to storage
-                await new Promise((resolve) =>
-                  chrome.storage.local.set({ 
-                    godamToken: newToken,
-                    godamRefreshToken: tokenData.refresh_token || godamRefreshToken,
-                    godamTokenExpiration: expirationTime
-                  }, () => resolve())
-                );
-
-                resolve(newToken);
-              } catch (error) {
-                // If refresh fails, try to sign in again
-                const newToken = await signInGoDAM();
-                if (!newToken) {
-                  reject(new Error("GoDAM token refresh and sign-in failed"));
-                }
-                resolve(newToken);
-              }
-            } else {
-              // Token is still valid
-              resolve(godamToken);
-            }
-          }
-        }
-      });
-    });
-  }
-
-  return new Promise(async (resolve, reject) => {
     try {
-      // Get the access token from Chrome storage
-      let token = await getGoDAMAuthToken();
+        const token = await getGoDAMAuthToken();
+        const formData = new FormData();
+        formData.append('file', videoBlob, fileName);
 
-      if (!token) {
-        throw new Error("GoDAM sign-in failed");
-      }
-      
-      const formData = new FormData();
-      formData.append('file', videoBlob, fileName);
+        const uploadUrl = process.env.GODAM_UPLOAD_URL || 'https://upload.godam.io'; // Todo: Replace the option with production URL
 
-      const uploadUrl =  process.env.GODAM_UPLOAD_URL || 'https://upload.godam.io'; // Todo: Replace the option with production URL
+        const url = uploadUrl + '/upload-file';
 
-      const url = uploadUrl + '/upload-file';
+        const uploadResponse = await fetch(
+            url,
+            {
+                method: "POST",
+                headers: {
+                    Authorization: `Bearer ${token}`,
+                },
+                body: formData,
+            }
+        );
 
-      const uploadResponse = await fetch(
-        url,
-        {
-          method: "POST",
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-          body: formData,
+        let message = 'An error occurred while saving to GoDAM!';
+        if (!uploadResponse.ok) {
+            // error message
+            if (uploadResponse.status === 400) {
+                message = 'An error occurred while saving to GoDAM! <br> Looks like you are not logged in to GoDAM. Please log in again.';
+            } else {
+                message = 'An error occurred while saving to GoDAM! <br> Please try again later, if the problem persists, please contact <a href="https://app.godam.io/helpdesk/my-tickets" target="blank">support team</a>.';
+            }
+            throw new Error(message);
         }
-      );
 
-      let message = 'An error occurred while saving to GoDAM!';
-      if (!uploadResponse.ok) {
-        // error message
-        if (uploadResponse.status === 400) {
-          message = 'An error occurred while saving to GoDAM! <br> Looks like you are not logged in to GoDAM. Please log in again.';
-        } else {
-          message = 'An error occurred while saving to GoDAM! <br> Please try again later, if the problem persists, please contact <a href="https://app.godam.io/helpdesk/my-tickets" target="blank">support team</a>.';
-        }
-        throw new Error(message);
-      }
+        const responseData = await uploadResponse.json();
+        const videoName = responseData?.file_informations?.name;
 
-      const responseData = await uploadResponse.json();
-      const filename = responseData.filename;
-      const location = responseData.location;
-      const videoName = responseData?.file_informations?.name;
+        const baseURL = process.env.GODAM_BASE_URL || 'https://app.godam.io';
 
-      const baseURL = process.env.GODAM_BASE_URL || 'https://app.godam.io';
-    
-      sendResponse({ status: "ok", url: `${baseURL}/web/video/${videoName}` }); 
+        sendResponse({ status: "ok", url: `${baseURL}/web/video/${videoName}` });
 
-      resolve(`${baseURL}/${location}`);
     } catch (error) {
-      sendResponse({ status: "error", url: null, message: error.message });
-      reject(error);
+        sendResponse({ status: "error", url: null, message: error.message });
+
     }
-  });
+
+
+
 };
 
 export default saveToGoDAM;
